Trixi2Img API

Trixi2Img.ndim — Constant

ndims

Number of spatial dimensions (= 2).

Trixi2Img.convert — Method

convert(filename::AbstractString...;
        format=:png, variables=[], verbose=false, grid_lines=false,
        output_directory=".", nvisnodes=nothing, max_supported_level=11)

Convert two-dimensional Trixi-generated output files to image files (PNG or PDF).

Arguments

filename: One or more 2D solution/restart files generated by Trixi to create an image from. Filenames support file globbing, e.g., "solution*" to match all files starting with solution.

format: Output format for solution/restart files. Can be 'vtu' or 'vti'.

variables: Names of the variables to create images for. If empty, each variable found in the input file(s) will be plotted.

verbose: Set to true to enable verbose output.

grid_lines: Plot outline of elements. (warning: this is an expensive operation!)

output_directory: Output directory where generated files are stored.

nvisnodes: Number of visualization nodes per element (default: twice the number of DG nodes). A value of 0 (zero) uses the number of nodes in the DG elements.

max_supported_level: Maximum cell refinement level supported for plotting.

Examples

julia> Trixi2Img.convert("out/solution_000*.h5")
[...]